Added test and edge case for incorrect file path.

+++ lib/web_logs.rb
@@ -11,6 +11,7 @@ class WebLogs
   end
 
   def create_sites_hash(filename)
+    fail "File does not exist." unless File.file?(filename)
     File.open(filename).each do |ip_site_logs|
       sites = ip_site_logs.split(' ')
       @sitelist.key?(sites[0]) ?
